`ignore` should understand both positional and keyword arguments Fixes #4
